Updating model to use 'hintable' instead of 'possible'
A card can possibly be a suit and still be unhintable as that suit if the possibilities are narrowed enough by indirect information.

Ex:
card 1 is not part of any of four prior hints - red, green, blue, and white - therefore it is known to be yellow and does not require a hint. It should not be possible to apply a yellow hint in this case.

So the card is possibly yellow (100 % possibility) but the "hintable" colors should not include it.